pyproject.toml
[project]
name = "bird-travel-recommender"
version = "0.1.0"
description = "A PocketFlow-based birding travel planning system using eBird data and LLM-powered recommendations"
authors = [
    {name = "Bird Travel Recommender Team"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
keywords = ["birding", "travel", "ebird", "recommendations", "llm", "pocketflow"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering",
    "Topic :: Other/Nonlisted Topic",
]
dependencies = [
    "aiohttp>=3.12.13",
    "mcp>=1.4.3,<2.0.0",
    "openai>=1.86.0",
    "pocketflow>=0.0.2",
    "pytest>=8.4.0",
    "pytest-asyncio>=1.0.0",
    "pytest-mock>=3.14.1",
    "python-dotenv>=1.1.0",
    "PyJWT>=2.8.0",
    "requests>=2.32.4",
]

[project.urls]
Homepage = "https://github.com/jeffkiefer/Bird-Travel-Recommender"
Repository = "https://github.com/jeffkiefer/Bird-Travel-Recommender"
Documentation = "https://github.com/jeffkiefer/Bird-Travel-Recommender/tree/main/docs"

[project.scripts]
bird-travel-recommender = "bird_travel_recommender.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/bird_travel_recommender"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--verbose",
]

[tool.coverage.run]
source = ["src"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
]

[dependency-groups]
dev = [
    "ruff>=0.12.2",
    "ty>=0.0.1a14",
]
